==> code/01_prepare_jttp_data.ipynb <==
import pandas as pd
import os 
import psycopg2
import logging

absolutepath = os.path.abspath(os.getcwd())
fileDirectory = os.path.dirname(absolutepath)
parentDirectory = os.path.dirname(fileDirectory)
dataprocDirectory = os.path.join(fileDirectory, 'data/proc')
datarawDirectory = os.path.join(fileDirectory, 'data/raw')

# copy to the raw directory main_jttp_analysis/data/raw/batch_all_clean2101020_merge_ra_clean.csv

# for chinese academy of science scholars - set affiliation to be the secondary

df = pd.read_csv(os.path.join(datarawDirectory, 'batch_all_clean2101020_merge_ra_clean.csv'))

df.loc[df.hiring_unit_id==60019499,'affname'] = df.loc[df.hiring_unit_id==60019499,'affname_secondary']
df.loc[df.hiring_unit_id==60019499,'hiring_unit_id'] = df.loc[df.hiring_unit_id==60019499,'hiring_unit_id_secondary'] 

# dealing with the never returned
# create variable never_returned
# use shaoyu's list
cp /cluster/work/lawecon/Work/dcai/jiycai/thousand_talent/thousand_talent/main_jttp_analysis/data/raw/renegers_returnee.xlsx /cluster/work/lawecon/Work/dcai/jiycai/thousand_talent/thousand_talent/scripts/10152022_iteration_peer_effect/data/raw

cp /cluster/work/lawecon/Work/dcai/jiycai/thousand_talent/thousand_talent/main_jttp_analysis/data/raw/renegers_nonreturnee.xlsx /cluster/work/lawecon/Work/dcai/jiycai/thousand_talent/thousand_talent/scripts/10152022_iteration_peer_effect/data/raw

df_returnee = pd.read_csv(os.path.join(datarawDirectory, 'renegers_returnee.xlsx'))

reneger = set(df_returnee.identifier)

df['never_returned'] = df.identifier.apply(lambda x: int(x in reneger))

df.to_csv(os.path.join(dataprocDirectory, 'jttp_data_cleaned.csv'),index=False)
