==> code/02_treatment_time_and_ever_treated_schools.ipynb <==
import pandas as pd
import os 
import psycopg2
import logging
import ast
absolutepath = os.path.abspath(os.getcwd())
fileDirectory = os.path.dirname(absolutepath)
parentDirectory = os.path.dirname(fileDirectory)
dataprocDirectory = os.path.join(fileDirectory, 'data/proc')
datarawDirectory = os.path.join(fileDirectory, 'data/raw')

# Read the cleaned data
jttp_cleaned = pd.read_csv(os.path.join(dataprocDirectory, 'jttp_data_cleaned.csv'))

# Create 'year' column by copying 'recruit_year'
jttp_cleaned['year'] = jttp_cleaned['recruit_year']

# Use ast.literal_eval to safely evaluate string representations of lists
jttp_cleaned['author_id_eval'] = [
    ast.literal_eval(v) if isinstance(v, str) else v  # Apply only if the value is a string
    for v in jttp_cleaned['author_id'].values
]

# Drop rows with NaN values in 'author_id_eval'
jttp_cleaned = jttp_cleaned.dropna(subset=['author_id_eval'])

# Extract the first element from each evaluated list (i.e., the first author_id)
jttp_cleaned['JTTP_authid'] = [v[0] for v in jttp_cleaned['author_id_eval'].values]


## merge in subfield
### copying the field file
cp /cluster/work/lawecon/Work/dcai/jiycai/thousand_talent/thousand_talent/scholar_fields.csv /cluster/work/lawecon/Work/dcai/jiycai/thousand_talent/thousand_talent/scripts/10152022_iteration_peer_effect/data/raw

scholar_field = pd.read_csv(os.path.join(datarawDirectory, 'scholar_fields.csv'))

jttp_cleaned = jttp_cleaned.merge(scholar_field, how = 'left', left_on = ['JTTP_authid'], right_on = ['authid'])

del scholar_field

# create treatment timing dataframe
## each row is (aff,2 digit field, 4 digit field, year of first jttp, fraction never returned among first impact, num jttp joining among first impact,)
## merge in subfield

groupby_variables_4digit = ['hiring_unit_id','subfield_most_frequent',
                            'subfield_most_frequent_two_digit']
outcome_variables_4digit = ['hiring_unit_id','subfield_most_frequent',
                            'subfield_most_frequent_two_digit',
                            'year','never_returned','JTTP_authid']
jttp_cleaned['first_impact_4digit_time'] = jttp_cleaned[outcome_variables_4digit].\
groupby(groupby_variables_4digit).transform('min')[['year']]
afid_subfield_timing = jttp_cleaned.loc[jttp_cleaned.year == jttp_cleaned.first_impact_4digit_time,
                                        outcome_variables_4digit].groupby(groupby_variables_4digit).\
agg({'year':'min','never_returned':'mean','JTTP_authid':'nunique'}).reset_index()


afid_subfield_timing['reneger_treatment_4digit'] = 1*(afid_subfield_timing['never_returned']==1)

afid_subfield_timing.never_returned.value_counts()

afid_subfield_timing.JTTP_authid.value_counts()

afid_subfield_timing[afid_subfield_timing.JTTP_authid == 5]

# what are the aff X 4 digit with 5 incoming at first impact?
## Beijing Computational Science Research Center
## Physical Sciences; Physics and Astronomy; Condensed Matter Physics;3104
### select * from scopus_affiliation where afid = 60104686;

## Shanghai Institute for Biological Sciences Chinese Academy of Sciences
## Life Sciences;Agricultural and Biological Sciences;Plant Science; 1110
### select * from scopus_affiliation where afid = 60007324;


#### https://github.com/plreyes/Scopus/blob/master/ASJC%20Codes%20with%20levels.csv

groupby_variables_2digit = ['hiring_unit_id',
                            'subfield_most_frequent_two_digit']
outcome_variables_2digit = ['hiring_unit_id',
                            'subfield_most_frequent_two_digit',
                            'year','never_returned','JTTP_authid']
jttp_cleaned['first_impact_2digit_time'] = jttp_cleaned[outcome_variables_2digit].\
groupby(groupby_variables_2digit).transform('min')[['year']]
afid_subfield_timing_2digit = jttp_cleaned.loc[jttp_cleaned.year == jttp_cleaned.first_impact_2digit_time,\
                                               outcome_variables_2digit].groupby(groupby_variables_2digit).\
agg({'year':'min','never_returned':'mean','JTTP_authid':'nunique'}).reset_index()

afid_subfield_timing_2digit['reneger_treatment_2digit'] = 1*(afid_subfield_timing_2digit['never_returned']==1)

afid_subfield_timing_2digit.never_returned.value_counts()

afid_subfield_timing_2digit.JTTP_authid.value_counts()

# save treatment timing files

afid_subfield_timing.to_csv(os.path.join(dataprocDirectory, 'afid_subfield_timing_4digit.csv'),index=False)
afid_subfield_timing_2digit.to_csv(os.path.join(dataprocDirectory, 'afid_subfield_timing_2digit.csv'),index=False)

# create list of ever treated schools

ever_treated_schools = jttp_cleaned.hiring_unit_id.unique()

# update the jttp_school column in scopus_affiliation

UPDATE scopus_affiliation
SET jttp_school = 0;

# create query

condition = " or ".join([f"afid = {afid}" for afid in ever_treated_schools])
query = f"""UPDATE scopus_affiliation SET jttp_school = 1 WHERE ({condition});"""

query

# create list of ever treated fields

ever_treated_fields = jttp_cleaned.subfield_most_frequent_two_digit.unique()

# update the jttp_field column in scopus_authors_aff_year_filled_chinese

ALTER TABLE scopus_authors_aff_year_filled_chinese ADD COLUMN jttp_field float;
UPDATE scopus_authors_aff_year_filled_chinese
SET jttp_field = 0;

condition = " or ".join([f"subfield_most_frequent_two_digit = '{field}'" for field in ever_treated_fields])
query = f"""UPDATE scopus_authors_aff_year_filled_chinese SET jttp_field = 1 WHERE ({condition});"""

query

jttp_cleaned.subfield_most_frequent_two_digit.count()
