==> code/03_create_dataset_candidate_pool_unrestricted.ipynb <==
import pandas as pd
import os
absolutepath = os.path.abspath(os.getcwd())
fileDirectory = os.path.dirname(absolutepath)
parentDirectory = os.path.dirname(fileDirectory)
dataprocDirectory = os.path.join(fileDirectory, 'data/proc')
datarawDirectory = os.path.join(fileDirectory, 'data/raw')

# the potential pool of scholars is
# all scholars that satisfy a), b), and c):
## a) in a two digit subfield that is ever treated across all cohorts
## b) in a jttp receiving school
## c) active at any time between 2010 and 2016

import psycopg2
import logging

def return_candiate_pool():
    # each row contains a authid X afid spell
    sql_text = f"""
    select authid,scopus_authors_aff_year_filled_chinese.afid,
    subfield_most_frequent,
    subfield_most_frequent_two_digit, 
    max(year) as afid_max_year, 
    min(year) as afid_min_year
    from scopus_authors_aff_year_filled_chinese
    left join scopus_affiliation on 
    scopus_authors_aff_year_filled_chinese.afid = scopus_affiliation.afid
    where scopus_affiliation.jttp_school = 1 and
    jttp_field = 1
    group by authid, 
    scopus_authors_aff_year_filled_chinese.afid, 
    subfield_most_frequent, 
    subfield_most_frequent_two_digit;
    """
    
    con = psycopg2.connect(host='id-hdb-psgr-cp7.ethz.ch',  
                      dbname='led', 
                      user='lixiang',
                      password='Wmhzgjwmhxgj99')

    cur = con.cursor()
    cur.execute(sql_text)
    rows = cur.fetchall()
    con.close()
    columns = ['authid','afid',
               'subfield_most_frequent',
               'subfield_most_frequent_two_digit',
               'afid_max_year',
               'afid_min_year']
    data = pd.DataFrame(rows,columns=columns)
    #keep all spells that terminated after 2011 and started before 2017 
    # fix - needs to be min year <= 2017
    data = data[((data.afid_max_year > 2011)&(data.afid_min_year < 2017))]
    data['spell_length'] = data['afid_max_year'] - data['afid_min_year'] + 1
    return data

data = return_candiate_pool()

data.spell_length.value_counts()

# keeping only :
# sum of spells at any jttp insitution larger than 1

data['total_spell'] = data[['authid','spell_length']].groupby(['authid']).transform('sum')[['spell_length']]

data.total_spell.value_counts()

data = data[data.total_spell>1]

data.to_csv(os.path.join(dataprocDirectory, 'author_pool_unrestricted.csv'),index=False)
