==> code/05_determine_treament_status_stacked.ipynb <==
import pandas as pd
import os
absolutepath = os.path.abspath(os.getcwd())
fileDirectory = os.path.dirname(absolutepath)
parentDirectory = os.path.dirname(fileDirectory)
dataprocDirectory = os.path.join(fileDirectory, 'data/proc')
datarawDirectory = os.path.join(fileDirectory, 'data/raw')

# load spells
# load treament timing

spells = pd.read_csv(os.path.join(dataprocDirectory, 'author_pool.csv'))

afid_subfield_timing_2digit = pd.read_csv(os.path.join(dataprocDirectory, \
                                                       'afid_subfield_timing_2digit.csv'))
afid_subfield_timing_2digit.columns = ['afid', 'subfield_most_frequent_two_digit', 'year',
       'never_returned', 'JTTP_authid', 'reneger_treatment_2digit']
afid_subfield_timing_4digit = pd.read_csv(os.path.join(dataprocDirectory, \
                                                       'afid_subfield_timing_4digit.csv'))
afid_subfield_timing_4digit.columns = ['afid', 'subfield_most_frequent',
       'subfield_most_frequent_two_digit', 'year', 'never_returned',
       'JTTP_authid', 'reneger_treatment_4digit']

# merge spell and treatment timing
# each row is (author X potential treament timing)
# 1. define presence
# presence if jttp year <  max year 
# 2. define treatment
# treated if jttp year <  max year & jttp year >=  min year + 1 

spellsXtiming_2digit = spells.merge(afid_subfield_timing_2digit, 
                                    how = 'outer',
                                    on = ['afid','subfield_most_frequent_two_digit'])
spellsXtiming_2digit = spellsXtiming_2digit.dropna(subset = ['authid'])

# presence 
# 1 : present 
spellsXtiming_2digit.loc[(spellsXtiming_2digit.year<\
                          spellsXtiming_2digit.afid_max_year),
                         'present_2digit'] = 1
# treatment 
# 1 : treated 
spellsXtiming_2digit.loc[(spellsXtiming_2digit.year>=\
                          spellsXtiming_2digit.afid_min_year+1)&\
                         (spellsXtiming_2digit.year<\
                          spellsXtiming_2digit.afid_max_year),
                         'treatment_2digit'] = 1
# 2 : join after treatment 
spellsXtiming_2digit.loc[(spellsXtiming_2digit.year<\
                          spellsXtiming_2digit.afid_min_year+1)&\
                         (spellsXtiming_2digit.year<\
                          spellsXtiming_2digit.afid_max_year),
                         'treatment_2digit'] = 2
print(spellsXtiming_2digit.shape)

spellsXtiming_2digit.treatment_2digit.value_counts()

## merge in 4 digit treament status

afid_subfield_timing_4digit['four_digit_match'] = 1
digit_vars = ['afid','year','subfield_most_frequent',
              'subfield_most_frequent_two_digit','four_digit_match']
spellsXtiming_2digit = spellsXtiming_2digit.merge(afid_subfield_timing_4digit[digit_vars],
                                                  how = 'left',
                                                  on = ['afid','year',
                                                        'subfield_most_frequent',
                                                        'subfield_most_frequent_two_digit'])

# for those treated
# among all possible treatment dates
# the earliest one is the assigned

groupbyvar = ['authid','subfield_most_frequent','subfield_most_frequent_two_digit']
groupbyvar_outcome = groupbyvar+['year','reneger_treatment_2digit','JTTP_authid','four_digit_match']
spellsXtiming_2digit_treated = spellsXtiming_2digit[spellsXtiming_2digit.treatment_2digit==1].copy()
spellsXtiming_2digit_treated['min_treat_year'] = spellsXtiming_2digit_treated.groupby(groupbyvar).\
transform('min')[['year']]
spellsXtiming_2digit_authid = spellsXtiming_2digit_treated.loc[spellsXtiming_2digit_treated.year==\
                                                       spellsXtiming_2digit_treated.min_treat_year,\
                                                       groupbyvar_outcome].groupby(groupbyvar).\
agg({'year':'min','reneger_treatment_2digit':'min',
     'JTTP_authid':'sum','four_digit_match':'max'}).reset_index()

spellsXtiming_2digit_authid.columns = ['authid', 'subfield_most_frequent', 
                                       'subfield_most_frequent_two_digit',
                                       'treat_year','reneger_treatment_2digit', 
                                       'jttp_count','four_digit_match']

# for those present but not treated - 
# ie those who join after jttp arrival
# label them

all_joiners = set(spellsXtiming_2digit[spellsXtiming_2digit.treatment_2digit == 2].authid)

len(all_joiners)

all_treated = set(spellsXtiming_2digit[spellsXtiming_2digit.treatment_2digit == 1].authid)

len(all_treated)

only_joiner = all_joiners - all_treated

len(only_joiner)

spellsXtiming_2digit_authid.count()

# combine to get authid level dataset

authid_level_dataset = spells[['authid','subfield_most_frequent','subfield_most_frequent_two_digit']].drop_duplicates()

authid_level_dataset = authid_level_dataset.merge(spellsXtiming_2digit_authid,
                                                  how = 'left',
                                                  on = ['authid',
                                                        'subfield_most_frequent',
                                                        'subfield_most_frequent_two_digit'])

authid_level_dataset['only_joiner'] = authid_level_dataset.authid.map(lambda x : 1*(x in only_joiner))

authid_level_dataset.shape

# authid x cohort dataset
# each cohort includes
# 1. those treated that year
# 2. those who are active the year before at one of the affiliations

spells_authid_first_last_year = spells.copy()
spells_authid_first_last_year['spell_min_year'] = spells_authid_first_last_year[['authid','afid_min_year']].groupby('authid')['afid_min_year'].transform('min')
spells_authid_first_last_year['spell_max_year'] = spells_authid_first_last_year[['authid','afid_max_year']].groupby('authid')['afid_max_year'].transform('max')
spells_authid_first_last_year = spells_authid_first_last_year[['authid','spell_min_year','spell_max_year']].drop_duplicates()

authid_level_dataset = authid_level_dataset.merge(spells_authid_first_last_year, how = 'left', on = 'authid')

authid_level_dataset

df_by_cohort_list = []
for c in [2011,2012,2013,2015,2016,2017]:
    print(c)
    print(authid_level_dataset.shape)
    cohort_c = authid_level_dataset[(authid_level_dataset.spell_min_year+1 <= c)&
                                    (authid_level_dataset.spell_max_year > c)].copy()
    #within cohort identify
    cohort_c['treatment_status'] = 0
    #treated in this cohort
    cohort_c.loc[cohort_c.treat_year==c,'treatment_status'] = 1
    #already treated or not yet treated
    cohort_c.loc[(cohort_c.treat_year<c)|(cohort_c.treat_year>c),'treatment_status'] = 2
    #never treated
    cohort_c.loc[pd.isna(cohort_c.treat_year),'treatment_status'] = 3
    print(cohort_c.treatment_status.value_counts())
    cohort_c['cohort'] = c
    print(cohort_c.shape)
    df_by_cohort_list.append(cohort_c)

df_by_cohort = pd.concat(df_by_cohort_list)

# drop jttp themselves

import ast

jttp_authid_list = []
for l in jttp_df.author_id.values:
    try:
        # Ensure that l is a string before attempting to eval
        if isinstance(l, str):
            jttp_authid_list.extend(ast.literal_eval(l))
        else:
            print(f"Skipping non-string value: {l}")
    except (ValueError, SyntaxError) as e:
        print(f"Error evaluating: {l}. Error: {e}")

jttp_authid_set = set(jttp_authid_list)
df_by_cohort['is_jttp'] = df_by_cohort.authid.apply(lambda x: 1 * (x in jttp_authid_set))


df_by_cohort = df_by_cohort[df_by_cohort.is_jttp!=1]

df_by_cohort.to_csv(os.path.join(dataprocDirectory, 'author_treatment_assignment_stacked.csv'),index=False)
