==> code/07_create_affiliation_vectors.ipynb <==
import pandas as pd
import os
absolutepath = os.path.abspath(os.getcwd())
fileDirectory = os.path.dirname(absolutepath)
parentDirectory = os.path.dirname(fileDirectory)
dataprocDirectory = os.path.join(fileDirectory, 'data/proc')
datarawDirectory = os.path.join(fileDirectory, 'data/raw')

import psycopg2
import logging

def return_candiate_pool_affiliation_history():
    # each row contains a authid X afid spell
    sql_text = f"""
    select scopus_authors_aff_year_filled_chinese.authid,scopus_authors_aff_year_filled_chinese.afid,
    subfield_most_frequent,
    subfield_most_frequent_two_digit,
    scopus_affiliation.jttp_school,
    year
    from scopus_authors_aff_year_filled_chinese
    left join scopus_affiliation on 
    scopus_authors_aff_year_filled_chinese.afid = scopus_affiliation.afid
    left join all_authors_in_restriction_jttp on 
    scopus_authors_aff_year_filled_chinese.authid = all_authors_in_restriction_jttp.authid
    where all_authors_in_restriction_jttp.restriction = 1;
    """
    
    con = psycopg2.connect(host='id-hdb-psgr-cp7.ethz.ch',  
                      dbname='led', 
                      user='jiycai')

    cur = con.cursor()
    cur.execute(sql_text)
    rows = cur.fetchall()
    con.close()
    columns = ['authid','afid',
               'subfield_most_frequent',
               'subfield_most_frequent_two_digit',
               'jttp_school',
               'year']
    data = pd.DataFrame(rows,columns=columns)
    return data

data = return_candiate_pool_affiliation_history()

data.shape

# check for gaps in affiliation size

data['max_year'] = data[['afid','authid','year']].groupby(['afid','authid']).transform('max')[['year']]
data['min_year'] = data[['afid','authid','year']].groupby(['afid','authid']).transform('min')[['year']]
data['spell_length'] = data.max_year - data.min_year + 1 

len(data[data.spell_length>1].afid.unique())

groupby_var = ['afid','subfield_most_frequent_two_digit']
groupby_var_outcome = groupby_var + ['authid']
data_affXafield_size = data[groupby_var_outcome].groupby(groupby_var).nunique().reset_index()

data_affXafield_size.columns = ['afid','subfield_most_frequent_two_digit','affXfield_size']

data_affXafield_size = data_affXafield_size.sort_values('affXfield_size')

data = data.merge(data_affXafield_size, how = 'left', on = groupby_var)

groupby_var = ['afid']
groupby_var_outcome = groupby_var + ['authid']
data_aff_size = data[groupby_var_outcome].groupby(groupby_var).nunique().reset_index()

data_aff_size = data_aff_size.sort_values('authid')

data_aff_size.columns = ['afid','afid_size']

data = data.merge(data_aff_size, how = 'left', on = 'afid')

data['afid_masked'] = data['afid']
data.loc[data.afid_size<100,'afid_masked'] = 'masked'

data.to_csv(os.path.join(dataprocDirectory, 'author_pool_affiliation_history.csv'),index=False)

# create affiliation list for each year

groupby_var = ['authid','year']
groupby_var_outcome = groupby_var + ['afid_masked']
data_vec = data[groupby_var_outcome].groupby(groupby_var).agg(['unique','nunique']).reset_index()
data_vec.columns = ['_'.join(col) for col in data_vec.columns]

data_vec.to_csv(os.path.join(dataprocDirectory, 'author_pool_affiliation_vectors_list_form.csv'),index=False)
