==> code/07b_create_affiliation_by_cohort.ipynb <==
import pandas as pd
import os
absolutepath = os.path.abspath(os.getcwd())
fileDirectory = os.path.dirname(absolutepath)
parentDirectory = os.path.dirname(fileDirectory)
dataprocDirectory = os.path.join(fileDirectory, 'data/proc')
datarawDirectory = os.path.join(fileDirectory, 'data/raw')

# load affiliation list for each year

data_vec = pd.read_csv(os.path.join(dataprocDirectory, 'author_pool_affiliation_vectors_list_form.csv'))

data_vec.afid_masked_nunique.value_counts()

data_vec.head()

5585374/7393644

# processs

def af_string_filter(af_string):
    af_list = af_string.replace("[",'').replace("]",'').split(' ')
    return af_list, len(af_list)

filtered_af = data_vec.afid_masked_unique.map(af_string_filter)

data_vec = data_vec.join(pd.DataFrame.from_records(filtered_af,
             columns = ['afid_masked_unique_filtered','afid_masked_nunique_filtered']))

data_vec_filtered_aff = data_vec[['authid_','year_','afid_masked_unique_filtered','afid_masked_nunique_filtered']]
data_vec_filtered_aff.columns = ['authid','cohort','afid_masked_unique_filtered','afid_masked_nunique_filtered']

# merge with panel and check fraction with only 1 af

balanced_dataset = pd.read_csv(os.path.join(dataprocDirectory,
                                                   'balanced_panel_wo_affiliation_never_treated_control_group.csv'))

balanced_dataset_authid_cohort = balanced_dataset[['authid','cohort']].drop_duplicates()

balanced_dataset_authid_cohort = balanced_dataset_authid_cohort.merge(data_vec_filtered_aff,
                                                                     how = 'left',
                                                                     on = ['authid','cohort'])

balanced_dataset_authid_cohort.afid_masked_nunique_filtered.value_counts()

balanced_dataset_authid_cohort.shape

685114/1051731

balanced_dataset_authid_cohort

balanced_dataset_authid_cohort.to_csv(os.path.join(dataprocDirectory, 'author_cohort_jttp_af.csv'),
                                      index=False)
