==> code/07c_create_baseline_reg_dataset_log.ipynb <==
import pandas as pd
import os
absolutepath = os.path.abspath(os.getcwd())
fileDirectory = os.path.dirname(absolutepath)
parentDirectory = os.path.dirname(fileDirectory)
dataprocDirectory = os.path.join(fileDirectory, 'data/proc')
datarawDirectory = os.path.join(fileDirectory, 'data/raw')

balanced_dataset = pd.read_csv(os.path.join(dataprocDirectory,
                                                   'balanced_panel_wo_affiliation_never_treated_control_group.csv'))

balanced_dataset['relative_period'] = balanced_dataset['year'] - balanced_dataset['cohort']

balanced_dataset = balanced_dataset[balanced_dataset.relative_period>-7]

import numpy as np
main_outcomes = ['num_docs','frac_count_docs','if_top10','if_top50','if_lower50',
                 'sum_citescore','funded']
for o in main_outcomes:
    balanced_dataset[f"log_1plus_{o}"] = np.log1p(balanced_dataset[o]) * 100

balanced_dataset['ever_treated'] = 0
balanced_dataset.loc[(balanced_dataset.treatment_status==1),'ever_treated'] = 1
relative_periods = balanced_dataset.relative_period.unique()
# create time period X treatment vars
# all periods
for rp in relative_periods:
    balanced_dataset[f"relative_period{rp}Xtreated"] = balanced_dataset.ever_treated*\
    (balanced_dataset.relative_period==rp)

balanced_dataset["postXtreated"] = balanced_dataset.ever_treated*\
                                          (balanced_dataset.relative_period>=0)

balanced_dataset["post0to3Xtreated"] = balanced_dataset.ever_treated*\
                                            ((balanced_dataset.relative_period>=0)&\
                                             (balanced_dataset.relative_period<4))
balanced_dataset["post4to8Xtreated"] = balanced_dataset.ever_treated*\
                                             (balanced_dataset.relative_period>=4)

cohort_af = pd.read_csv((os.path.join(dataprocDirectory,
                    'author_cohort_jttp_af.csv')))

balanced_dataset = balanced_dataset.merge(cohort_af, how = 'left', on =['authid','cohort'])

balanced_dataset.count()

# drop only joiners

balanced_dataset = balanced_dataset[balanced_dataset.only_joiner!=1]

balanced_dataset.shape

balanced_dataset.to_csv(os.path.join(dataprocDirectory,'balanced_panel_never_treated_control_group_log.csv'),index=False)
